Changed TV Glitch time to be based on seconds, not milliseconds
This will break existing code. Trying to standardize this across effects.
Simplex works on seconds, and it makes more sense this way since a) it's
easier to think about, b) video currentTime is in seconds and c) some
weird noise functions go a bit crazy with very, very high numbers.