Change canvas output to premultiplied alpha (#88)
A workaround for broken/missing support for the `premultipliedAlpha` WebGL context option in Safari and Internet Explorer. This is not ideal, because it changes the output values of `readPixels` on WebGL target nodes. But it appears to be the only way to make transparent backgrounds work in all browsers.

According to the WebGL spec, support for the option is mandatory, so there is no way to feature-detect whether it's working in a given browser. So the workaround needs to apply everywhere.